--- src/src/pages/page1.js
@@ -1,29 +1,0 @@
-import React from 'react';
-  
-const Page1 = () => {
-  return (
-    <div
-    >
-      <h1>This is page 1.</h1>
-      <h1>Lorem ipsum dolor sit amet</h1>
-      <h1>consectetur adipiscing elit, sed do</h1>
-      <h1>eiusmod tempor incididunt ut labore</h1>
-      <h1>fugiat nulla pariatur</h1>
-      <br></br><br></br><br></br><br></br>
-      <br></br>
-      <br></br>
-      <br></br>
-      <h1>fugiat nulla pariatur</h1>
-      <br></br><br></br><br></br><br></br>
-      <br></br><br></br><br></br>
-      <h1>fugiat nulla pariatur</h1>
-      <h1>fugiat nulla pariatur</h1>
-      <br></br><br></br><br></br><br></br>
-      <br></br><br></br><br></br>
-      <h1>fugiat nulla pariatur</h1>
-      
-    </div>
-  );
-};
-  
-export default Page1;
