+++ src/src/pages/signin.js
@@ -1,0 +1,18 @@
+import React from 'react';
+  
+const SignIn = () => {
+  return (
+    <div
+      style={{
+        display: 'flex',
+        justifyContent: 'Right',
+        alignItems: 'Right',
+        height: '100vh'
+      }}
+    >
+      <h1>Sign in page   yay</h1>
+    </div>
+  );
+};
+  
+export default SignIn;
